Replace pickle checkpoint with YAML + preset-based config reload
Save: config.yaml (logging only) + train_state.yaml (scalars + alg metadata)
Load: re-run preset via preset_module → get_config(), no YAML-to-config merge

- base_runner: save config.yaml + train_state.yaml instead of metadata.pkl
- checkpoint.py: load_config_from_checkpoint() imports preset and calls get_config()
- All runners: use load_config_from_checkpoint instead of configs_from_dict
- All sim_initializers: use load_config_from_checkpoint for same-sim eval
- evaluator: disable interval events via setattr(None) instead of list filter
- All algorithms + runners: fix uint32 overflow for JAX keys loaded from YAML

+++ JaxRLWorld/rlworld/rl/algorithms/fast_td3/fast_td3.py
@@ -607,7 +607,7 @@ class FastTD3(OffPolicyAlgorithm):
             actor_opt_state=actor_opt_state,
             critic_opt_state=critic_opt_state,
             noise_scales=noise_scales,
-            key=jnp.array(metadata["alg_key"]),
+            key=jnp.array(metadata["alg_key"], dtype=jnp.uint32),
         )
 
         self.actor_lr = metadata.get("actor_lr", self.actor_lr)
