+++ JaxRLWorld/rlworld/rl/algorithms/ppo_dr3/ppo_dr3.py
@@ -260,7 +260,7 @@ class PPODR3(PPO):
         self.train_state = PPOTrainState(
             model=new_model,
             opt_state=new_opt_state,
-            key=jnp.array(metadata["alg_key"]),
+            key=jnp.array(metadata["alg_key"], dtype=jnp.uint32),
         )
 
         self.actor_lr = metadata.get("actor_lr", self.actor_lr)
