+++ JaxRLWorld/rlworld/rl/algorithms/td3/td3.py
@@ -514,7 +514,7 @@ class TD3(OffPolicyAlgorithm):
             target_critic2_static=new_critic2_static,
             actor_opt_state=actor_opt_state,
             critic_opt_state=critic_opt_state,
-            key=jnp.array(metadata["alg_key"]),
+            key=jnp.array(metadata["alg_key"], dtype=jnp.uint32),
         )
 
         self.actor_lr = metadata.get("actor_lr", self.actor_lr)
